requirements.txt
aibread>=0.1.0
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
numpy>=1.24.0

